requirements.txt
discord.py>=2.6.4
python-dotenv>=1.0.0
appwrite==11.1.0

